@@ -34,12 +34,13 @@ export default {
         'zIndex': -1
     },
     'panel': {
+        'backgroundColor': '#fff',
         'borderRadius': '5px',
         'boxSizing': 'border-box',
         'boxShadow': '0 2px 4px rgba(0, 0, 0, 0.3)',
         'transform': 'translate3d(0, 0, 0)',
         'transition': 'transform 500ms cubic-bezier(0, 0, 0.25, 1), opacity 500ms cubic-bezier(0, 0, 0.25, 1)',
-        'zIndex': 10002
+        'zIndex': 10002,
     },
     'panelHidden': {
         'transform': 'translate3d(0, -100px, 0)',
